chore(node): rename fiber
