+++ demo/midgard-node/src/fibers/index.ts
@@ -1,5 +1,5 @@
 export * from "./block-commitment.js";
-export * from "./submit-blocks.js";
+export * from "./block-submission.js";
 export * from "./sync-user-events.js";
 export * from "./merge.js";
 export * from "./monitor-mempool.js";
